door/requirements.txt
flask
flask-socketio
eventlet
pyserial
python-pptx
fpdf2
